Drop Linear::evaluate_bound
